完善conn manage部分

+++ pkg/ciface/iconnmanage.go
@@ -3,7 +3,7 @@ package ciface
 type IConnManager interface {
 	Add(conn IConnection)                   //添加链接
 	Remove(conn IConnection)                //删除连接
-	Get(connID uint32) (IConnection, error) //利用ConnID获取链接
+	Get(connID uint64) (IConnection, error) //利用ConnID获取链接
 	Len() int                               //获取当前连接
 	ClearConn()                             //删除并停止所有链接
 }
